On application mount, check/create metadata branch
